Pages dir moved to src from frontend dir
