+++ frontend/src/main.jsx
@@ -6,14 +6,14 @@ import { Provider } from 'react-redux'
 import App from './App.jsx'
 import 'bootstrap/dist/css/bootstrap.min.css'
 import './index.css'
-import Home from '../pages/Home.jsx'
-import LoginPage from '../pages/LoginPage.jsx'
-import RegisterPage from '../pages/RegisterPage.jsx'
-import ProfilePage from '../pages/ProfilePage.jsx'
+import Home from './pages/Home.jsx'
+import LoginPage from './pages/LoginPage.jsx'
+import RegisterPage from './pages/RegisterPage.jsx'
+import ProfilePage from './pages/ProfilePage.jsx'
 import PrivateRoute from './components/PrivateRoute.jsx'
-import Devices from '../pages/Devices.jsx'
-import AddDevice from '../pages/AddDevice.jsx'
-import Device from '../pages/Device.jsx'
+import Devices from './pages/Devices.jsx'
+import AddDevice from './pages/AddDevice.jsx'
+import Device from './pages/Device.jsx'
 
 // PAGES
 const router = createBrowserRouter(
@@ -21,9 +21,9 @@ const router = createBrowserRouter(
     <Route path='/' element={ <App /> }>
       <Route index={true} path='/' element={ <Home />} />
       <Route path='/login' element={ <LoginPage />} />
+        <Route path='/register' element={ <RegisterPage />} />
       {/* PRIVATE ROUTES */}
       <Route path='' element={<PrivateRoute />}>
-        <Route path='/register' element={ <RegisterPage />} />
         <Route path='/profile' element={ <ProfilePage />} />
         <Route path='/devices' element={ <Devices />} />
         <Route path='/devices/add' element={ <AddDevice />} />
