+++ frontend/src/pages/Devices.jsx
