+++ frontend/src/pages/LoginPage.jsx
@@ -3,11 +3,11 @@ import { useState, useEffect } from 'react'
 import { Link, useNavigate } from 'react-router-dom'
 import { Form, Button, Row, Col } from 'react-bootstrap'
 import { useDispatch, useSelector } from 'react-redux'
-import FormContainer from '../src/components/FormContainer'
-import { useLoginMutation } from '../src/slices/usersApiSlice'
-import { setCredentials } from '../src/slices/authSlice'
+import FormContainer from '../components/FormContainer'
+import { useLoginMutation } from '../slices/usersApiSlice'
+import { setCredentials } from '../slices/authSlice'
 import { toast } from 'react-toastify'
-import Loader from '../src/components/Loader'
+import Loader from '../components/Loader'
 
 const LoginPage = () => {
     const [email, setEmail] = useState('')
